electricity2030: pass entry values, split files
