+++ src/climatevision/generator/electricity2030/commune/energy_demand/calc_energy_demand.py
@@ -1,0 +1,17 @@
+# pyright: strict
+
+from ....refdata import Facts
+from ....utils import div
+
+from ...core.electricity2030_core import EnergyDemandWithCostFuel
+
+
+def calc_energy_demand(
+    facts: Facts, *, energy: float, energy_18: float, cost_fuel_per_MWh: str
+) -> EnergyDemandWithCostFuel:
+    d = EnergyDemandWithCostFuel()
+    d.energy = energy
+    d.cost_fuel_per_MWh = facts.fact(cost_fuel_per_MWh)
+    d.change_energy_MWh = energy - energy_18
+    d.change_energy_pct = div(d.change_energy_MWh, energy_18)
+    return d
